Release 1.1.2

+++ worker/src/main/java/io/github/hyshmily/hotkey/worker/WorkerProperties.java
@@ -19,10 +19,20 @@ import jakarta.validation.constraints.Min;
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+/**
+ * Configuration for the standalone HotKey Worker server.
+ *
+ * <p>Prefix: {@code hotkey.worker}.
+ *
+ * <p>Groups cover routing and sharding, AMQP exchange names, sliding-window
+ * parameters, threshold settings, state-machine timings, dynamic threshold
+ * adaptation, TopK pre-warm validation, and HeavyKeeper algorithm tuning.
+ */
 @Data
 @ConfigurationProperties(prefix = "hotkey.worker")
 public class WorkerProperties {
 
+  /** Routing configuration for sharding app-level report queues. */
   @Data
   public static class Routing {
 
@@ -31,6 +41,7 @@ public class WorkerProperties {
     private int shardIndex = 0;
   }
 
+  /** AMQP exchange names for report and broadcast messages. */
   @Data
   public static class Messaging {
 
@@ -38,6 +49,7 @@ public class WorkerProperties {
     private String broadcastExchange = "hotkey.broadcast.exchange";
   }
 
+  /** Sliding-window parameters for local QPS tracking. */
   @Data
   public static class SlidingWindow {
 
@@ -47,6 +59,7 @@ public class WorkerProperties {
     private int slices = 10;
   }
 
+  /** Absolute and ratio-based thresholds for HOT key classification. */
   @Data
   public static class Threshold {
 
@@ -54,6 +67,7 @@ public class WorkerProperties {
     private double hotThresholdRatio = 0.01;
   }
 
+  /** State-machine timing for HOT/COOL decision transitions. */
   @Data
   public static class StateMachine {
 
@@ -62,15 +76,17 @@ public class WorkerProperties {
     private long preCoolGraceMs = 5000;
   }
 
+  /** Dynamic threshold adaptation based on global QPS changes. */
   @Data
   public static class GlobalQpsDynamicThreshold {
 
     private double qpsChangeTolerance = 0.5;
-    private double learningPeriodMs = 30_000;
+    private long learningPeriodMs = 30_000;
     private double hotThresholdRatio = 0.01;
     private long recalculateIntervalMs = 60_000;
   }
 
+  /** Pre-warm validation before emitting HOT decisions. */
   @Data
   public static class TopKValidation {
 
@@ -79,6 +95,7 @@ public class WorkerProperties {
     private int preWarmMinAppearances = 2;
   }
 
+  /** HeavyKeeper TopK algorithm parameters for worker-side hot key detection. */
   @Data
   public static class HeavyKeeper {
 
@@ -99,16 +116,31 @@ public class WorkerProperties {
   private TopKValidation topKValidation = new TopKValidation();
   private HeavyKeeper heavyKeeper = new HeavyKeeper();
 
+  /**
+   * Number of sliding-window slices that fit within the CONFIRM duration.
+   *
+   * @return window count for the confirm phase
+   */
   public int getConfirmWindows() {
     double sliceMs = (double) slidingWindow.getDurationMs() / slidingWindow.getSlices();
     return (int) Math.ceil(stateMachine.getConfirmDurationMs() / sliceMs);
   }
 
+  /**
+   * Number of sliding-window slices that fit within the COOL duration.
+   *
+   * @return window count for the cool phase
+   */
   public int getCoolWindows() {
     double sliceMs = (double) slidingWindow.getDurationMs() / slidingWindow.getSlices();
     return (int) Math.ceil(stateMachine.getCoolDurationMs() / sliceMs);
   }
 
+  /**
+   * Number of sliding-window slices that fit within the pre-cool grace period.
+   *
+   * @return window count for the pre-cool grace phase
+   */
   public int getPreCoolGraceWindows() {
     double sliceMs = (double) slidingWindow.getDurationMs() / slidingWindow.getSlices();
     return (int) Math.ceil(stateMachine.getPreCoolGraceMs() / sliceMs);
